--- components/BackgroundImg.jsx
@@ -1,19 +1,0 @@
-import Image from "next/image";
-
-export default function BackgroundImg({ src }) {
-  return (
-    <div className="w-full relative overflow-hidden rounded-[20px] bg-background">
-      <div
-        className="w-full"
-        data-scroll
-        data-scroll-speed="-.2.5"
-      >
-        <Image
-          src={src}
-          alt="img"
-          className="w-full h-full object-cover"
-        />
-      </div>
-    </div>
-  );
-}
