+++ components/Curve/Curve.jsx
@@ -5,65 +5,65 @@ import { useEffect, useState } from "react";
 import { text, curve, translate } from "@/motion";
 
 const routes = {
-  "/": "Home",
-  "/services": "Features",
-  "/presentation": "Why Us",
-  "/ochi-team": "About Us",
-  "/insights": "FAQ",
-  "/contact": "Contact Us",
-  "/case": "Workiz Easy",
+	"/": "Home",
+	"/services": "Features",
+	"/presentation": "Why Us",
+	"/ochi-team": "About Us",
+	"/insights": "FAQ",
+	"/contact": "Contact Us",
+	"/case": "Workiz Easy",
 };
 
 const anim = (variants) => {
-  return {
-    variants,
-    initial: "initial",
-    animate: "enter",
-    exit: "exit",
-  };
+	return {
+		variants,
+		initial: "initial",
+		animate: "enter",
+		exit: "exit",
+	};
 };
 
 export default function Curve({ children, backgroundColor }) {
-  const router = useRouter();
-  const [dimensions, setDimensions] = useState({
-    width: null,
-    height: null,
-  });
+	const router = useRouter();
+	const [dimensions, setDimensions] = useState({
+		width: null,
+		height: null,
+	});
 
-  useEffect(() => {
-    function resize() {
-      setDimensions({
-        width: window.innerWidth,
-        height: window.innerHeight,
-      });
-    }
-    resize();
-    window.addEventListener("resize", resize);
-    return () => {
-      window.removeEventListener("resize", resize);
-    };
-  }, []);
+	useEffect(() => {
+		function resize() {
+			setDimensions({
+				width: window.innerWidth,
+				height: window.innerHeight,
+			});
+		}
+		resize();
+		window.addEventListener("resize", resize);
+		return () => {
+			window.removeEventListener("resize", resize);
+		};
+	}, []);
 
-  return (
-    <div style={{ backgroundColor }}>
-      <div
-        style={{ opacity: dimensions.width == null ? 1 : 0 }}
-        className="fixed h w-full pointer-events-none left-0 top-0 z-50 bg-black"
-      />
-      <motion.p
-        className="absolute left-1/2 top-[40%] text-white text-[50px] z-[60] -translate-x-1/2 text-center"
-        {...anim(text)}
-      >
-        {routes[router.route]}
-      </motion.p>
-      {dimensions.width != null && <SVG {...dimensions} />}
-      {children}
-    </div>
-  );
+	return (
+		<div style={{ backgroundColor }}>
+			<div
+				style={{ opacity: dimensions.width == null ? 1 : 0 }}
+				className="fixed h w-full pointer-events-none
+				 left-0 top-0 z-50 bg-black"
+			/>
+			<motion.p
+				className="absolute left-1/2 top-[40%] text-white text-[50px] z-[60] -translate-x-1/2 text-center"
+				{...anim(text)}>
+				{routes[router.route]}
+			</motion.p>
+			{dimensions.width != null && <SVG {...dimensions} />}
+			{children}
+		</div>
+	);
 }
 
 const SVG = ({ height, width }) => {
-  const initialPath = `
+	const initialPath = `
         M0 300 
         Q${width / 2} 0 ${width} 300
         L${width} ${height + 300}
@@ -71,7 +71,7 @@ const SVG = ({ height, width }) => {
         L0 0
     `;
 
-  const targetPath = `
+	const targetPath = `
         M0 300
         Q${width / 2} 0 ${width} 300
         L${width} ${height}
@@ -79,12 +79,12 @@ const SVG = ({ height, width }) => {
         L0 0
     `;
 
-  return (
-    <motion.svg
-      className="fixed h w-full pointer-events-none left-0 top-0 z-50"
-      {...anim(translate)}
-    >
-      <motion.path {...anim(curve(initialPath, targetPath))} />
-    </motion.svg>
-  );
+	return (
+		<motion.svg
+			className="fixed h w-full pointer-events-none
+				 left-0 top-0 z-50"
+			{...anim(translate)}>
+			<motion.path {...anim(curve(initialPath, targetPath))} />
+		</motion.svg>
+	);
 };
