+++ components/Marquee.tsx
@@ -1,0 +1,21 @@
+import { TMarqueeProps } from "@/types";
+import { TextMarquee } from "@/animation";
+
+export default function Marquee({ title, className }: TMarqueeProps) {
+	return (
+		<TextMarquee baseVelocity="0.7">
+			<h1
+  className={`font-FoundersGrotesk font-normal border-y border-[#ffffff55] uppercase text-white whitespace-nowrap tracking-[-5px] ${className}`}
+  style={{ backgroundColor: "#005A9C" }}>
+  {title} &nbsp;
+</h1>
+
+			<h1
+  className={`font-FoundersGrotesk font-normal border-y border-[#ffffff55] uppercase text-white whitespace-nowrap tracking-[-5px] ${className}`}
+  style={{ backgroundColor: "#005A9C" }}>
+  {title} &nbsp;
+</h1>
+
+		</TextMarquee>
+	);
+}
