--- components/MobileNav.jsx
@@ -1,78 +1,0 @@
-"use client";
-import Link from "next/link";
-import { useState } from "react";
-import { IoMdClose } from "react-icons/io";
-import { footernavbarItems } from "@/constants";
-import { HiOutlineMenuAlt4 } from "react-icons/hi";
-import { AnimatePresence, motion } from "framer-motion";
-
-export default function MobileNav() {
-  const [toggle, setToggle] = useState(false);
-
-  return (
-    <>
-      {/* ✅ Top navbar with text logo */}
-      <div 
-        className="w-full hidden justify-between items-center h-[8vh] padding-x sm:flex xm:flex md:flex"
-        style={{
-          backgroundColor: "#6495ED",
-          borderRadius: "0px 0px 25px 25px",
-          boxShadow: "0 4px 20px rgba(0, 0, 0, 0.15)",
-        }}
-      >
-        <Link href={"/"}>
-          <h1 className="text-2xl font-extrabold font-sans uppercase tracking-wide text-white">
-            Athena <span className="text-white">LMS</span>
-          </h1>
-        </Link>
-        <HiOutlineMenuAlt4
-          onClick={() => setToggle(true)}
-          className="text-3xl cursor-pointer text-white"
-        />
-      </div>
-
-      <AnimatePresence mode="wait">
-        {toggle && (
-          <motion.div
-            initial={{ y: "-100%" }}
-            animate={{ y: 0 }}
-            exit={{ y: "-100%" }}
-            transition={{ duration: 1, ease: [0.3, 0.86, 0.36, 0.95] }}
-            className="fixed top-0 bottom-0 right-0 z-50 w-full min-h-screen flex justify-end items-end flex-col"
-            style={{
-              backgroundColor: "#6495ED",
-              borderRadius: "0px 0px 0px 25px",
-            }}
-          >
-            {/* ✅ Mobile overlay with text logo */}
-            <div className="w-full flex justify-between items-center h-[8vh] border-b border-[#f1f1f155] padding-x">
-              <Link href={"/"}>
-                <h1 className="text-2xl font-extrabold font-sans uppercase tracking-wide text-white">
-                  Athena <span className="text-white">LMS</span>
-                </h1>
-              </Link>
-              <IoMdClose
-                onClick={() => setToggle(false)}
-                className="text-3xl cursor-pointer text-white"
-              />
-            </div>
-
-            {/* ✅ Menu items */}
-            <ul className="h-full w-full flex justify-center text-left flex-col gap-[10px] padding-x">
-              {footernavbarItems.map((item) => (
-                <Link
-                  href={item.href}
-                  key={item.id}
-                  onClick={() => setToggle(false)}
-                  className="text-[80px] leading-[67px] font-FoundersGrotesk uppercase font-bold tracking-[-.9] text-white"
-                >
-                  {item.title}
-                </Link>
-              ))}
-            </ul>
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </>
-  );
-}
